+++ app/src/main/java/architex/labs/coffeedrop/presentation/components/CoffeeFilterOption.kt
@@ -16,8 +16,6 @@
 package architex.labs.coffeedrop.presentation.components
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -26,16 +24,14 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.composed
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import architex.labs.coffeedrop.domain.models.sub_models.CoffeeType
 import architex.labs.coffeedrop.presentation.theme.CircularStd
 import architex.labs.coffeedrop.presentation.theme.Neutrals200
 import architex.labs.coffeedrop.presentation.theme.Primary
+import architex.labs.coffeedrop.presentation.utils.clickableNoRipple
 
 @Composable
 fun CoffeeFilterOption(
@@ -44,22 +40,6 @@ fun CoffeeFilterOption(
 	isActiveOption: Boolean,
 	updateSelectedCoffeeType: () -> Unit
 ) {
-	// Extending Modifier to disable ripple effect
-	fun Modifier.clickableNoRipple(
-		enabled: Boolean = true,
-		onClickLabel: String? = null,
-		role: Role? = null,
-		onClick: () -> Unit,
-	): Modifier = composed {
-		this.clickable(
-			interactionSource = remember { MutableInteractionSource() },
-			indication = null,
-			enabled = enabled,
-			onClickLabel = onClickLabel,
-			role = role,
-			onClick = onClick
-		)
-	}
 
 	Column(
 		modifier = modifier
