+++ app/src/main/res/values/strings.xml
@@ -21,6 +21,7 @@
     <string name="description_menu">Menu Icon</string>
     <string name="description_avatar">Avatar</string>
     <string name="description_search">Search Icon</string>
+    <string name="description_clear">Clear Icon</string>
     <string name="description_rating">Rating</string>
     <string name="description_add_button">Add Button</string>
     <string name="description_for_you_image">For You Image</string>
